Move large account.steps.ts into smaller files by responsibility.
